// models/run_data.ts
export interface RunData {
  schedule_index: number;
  pb: string;
  estimate: string;
  pb_seconds: number;
  estimate_seconds: number;
  actual_start_time: string;
  actual_end_time: string;
  actual_time_seconds: number;
  team_name: string;
  team_color: string;
  username: string;
  twitch: string;
  twitter: string;
  avatar_object_id: string;
  game_name: string;
  game_series: string;
  progress_unit: string;
  progress_max: string;
}